+++ src/gui/crystalModel.cpp
@@ -1,5 +1,4 @@
 #include "crystalModel.h"
-#include "../simulation/defaults.h"
 
 CrystalModel::CrystalModel(std::shared_ptr<HaloSim::CrystalPopulationRepository> crystalRepository, QWidget *parent)
     : QAbstractTableModel(parent),
@@ -9,7 +8,7 @@ CrystalModel::CrystalModel(std::shared_ptr<HaloSim::CrystalPopulationRepository>
 
 int CrystalModel::rowCount(const QModelIndex &parent) const
 {
-    return static_cast<int>(mCrystals->GetCount());
+    return static_cast<int>(mCrystals->getCount());
 }
 
 int CrystalModel::columnCount(const QModelIndex &parent) const
@@ -24,7 +23,7 @@ QVariant CrystalModel::data(const QModelIndex &index, int role) const
 
     auto col = index.column();
     auto row = index.row();
-    const HaloSim::CrystalPopulation &crystal = mCrystals->Get(row);
+    const HaloSim::CrystalPopulation &crystal = mCrystals->get(row);
 
     switch (col)
     {
@@ -45,7 +44,7 @@ QVariant CrystalModel::data(const QModelIndex &index, int role) const
     case 7:
         return crystal.rotationStd;
     case 8:
-        return mCrystals->GetWeight(row);
+        return mCrystals->getWeight(row);
     }
 
     return QVariant();
@@ -61,7 +60,7 @@ bool CrystalModel::setData(const QModelIndex &index, const QVariant &value, int 
 
     auto row = index.row();
     auto col = index.column();
-    HaloSim::CrystalPopulation &crystal = mCrystals->Get(row);
+    HaloSim::CrystalPopulation &crystal = mCrystals->get(row);
     switch (col)
     {
     case 0:
@@ -89,7 +88,7 @@ bool CrystalModel::setData(const QModelIndex &index, const QVariant &value, int 
         crystal.rotationStd = value.toFloat();
         break;
     case 8:
-        mCrystals->SetWeight(row, value.toUInt());
+        mCrystals->setWeight(row, value.toUInt());
         break;
     default:
         break;
@@ -105,21 +104,21 @@ Qt::ItemFlags CrystalModel::flags(const QModelIndex &index) const
     return Qt::ItemIsEditable | QAbstractTableModel::flags(index);
 }
 
-void CrystalModel::addRow()
+void CrystalModel::addRow(HaloSim::CrystalPopulationPreset preset)
 {
-    auto row = mCrystals->GetCount();
+    auto row = mCrystals->getCount();
     beginInsertRows(QModelIndex(), row, row);
-    mCrystals->AddDefault();
+    mCrystals->add(preset);
     endInsertRows();
 }
 
 bool CrystalModel::removeRow(int row)
 {
-    if (mCrystals->GetCount() <= 1)
+    if (mCrystals->getCount() <= 1)
         return false;
 
     beginRemoveRows(QModelIndex(), row, row);
-    mCrystals->Remove(row);
+    mCrystals->remove(row);
     endRemoveRows();
 
     return true;